Dev/authorization

+++ server/src/main/java/Commands/RemoveFirst.java
@@ -2,15 +2,15 @@ package Commands;
 
 import SharedModels.Response;
 import SharedUtility.ResponseStatus;
-import Utility.Command;
 import SharedUtility.CommandType;
+import Utility.OwnerCommand;
 
 import java.sql.Connection;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.Statement;
 import java.util.ArrayList;
 
-public class RemoveFirst extends Command {
+public class RemoveFirst extends OwnerCommand {
     private final Connection connection;
     public RemoveFirst(Connection connection) {
         super("remove_first", CommandType.REMOVE_FIRST);
@@ -18,15 +18,16 @@ public class RemoveFirst extends Command {
     }
 
     @Override
-    public Response execute(ArrayList<Object> args) {
+    public Response execute(ArrayList<Object> args, String username) {
         try {
-            Statement st = connection.createStatement();
-            ResultSet rs = st.executeQuery(
-                    """
+            PreparedStatement st = connection.prepareStatement("""
                             DELETE FROM MusicBand
                             WHERE id = (SELECT MIN(id) FROM MusicBand)
+                            AND creator = ?
                             RETURNING id;
                             """);
+            st.setString(1, username);
+            ResultSet rs = st.executeQuery();
             if (rs.next() && rs.getLong(1) > 0){
                 return new Response(ResponseStatus.OK, "Группа удалена.");
             }
